gateway/requirements.txt
# ---- Core ----
fastapi==0.114.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.4.0

# ---- Adapters / external calls ----
httpx==0.27.0   # асинхронні HTTP-запити (для OpenAI, інших сервісів)

# ---- Auth / Security ----
python-jose[cryptography]==3.3.0  # робота з JWT
passlib[bcrypt]==1.7.4            # хешування паролів (якщо буде локальна реєстрація)

# ---- Utils ----
python-dotenv==1.0.1              # читання .env (альтернатива pydantic-settings)
loguru==0.7.2                     # зручне логування

# ---- Dev / Tests ----
pytest==8.3.2
httpx==0.27.0                     # також використовується в тестах